requirements.txt
requests
sqlmodel
python-dotenv